Box2d (#1)
* Optimize CMakeLists.txt find patterns

* Refactor component and GOManager

* Make CMakeLists greate again

* Separate some src sub-directories

* Refactor RenderComponent

* Fix deltaTime

+++ core/GameObject.hpp
@@ -35,7 +35,7 @@ public:
 	}
 
 	void start() const;
-	void update(const float deltaTime) const;
+	void update(const int64_t deltaTime) const;
 
 	void destroy()
 	{
